Add files via upload
Modified the Spacehunt.html with a added link button for devs to access separate html page for the user to select and choose the state variables as needed. Also created the devOptions.html file that has the front end presentation for the dev to make the selections via the use of tab boxes. Will continue progress on this later where on the submit button it will return to the spacehunt.html page and populate the values with whatever input was given in the devOptions page"