Add files via upload
In the devOptions.html I updated the way the dev input is stored, instead of a string array it is now all concatenated into one large string with comma parses between so it can be saved to the local storage under the title 'Config'. I also created the close() function so that on submit press the page will go back to the SpaceHunt.html file. 

In the gameObjects.js file I created a if check at the beginning to see if the 'Config' is saved in the local storage and if it is it will pull the string out, split the string by the comma parses into a string array and then convert each into int values. I then clear out the local storage and set a bool var to true. Later on I plan to use that bool in the spaceship object to choose between what values to set the initial set up to be.