devOptions.html
Updated devOptions.html to store the radio button data correctly and accurately, also added a out of bounds check for current location based on what is inputted in for the game size. Set up splits on the game size and current location input to split the 2 values so i can apply them into the gameObjects.js correctly.