Add files via upload
Fixed up some formatting in the setdata() function in the devOptions.html file, set up a split method on the fixed wormhole section along with a check if condition to confirm that the wormhole would be in the map size specified by the dev. Set up all the ternary conditionals on the initial value sets so if a devconfig is set it will change to the config values along with wormhole and gamespace functions. I still need to modify nickolai's files for regular or immortal play and perform some more testing on these changes later.