Add WIP handlers for space station events
This includes the onclick event for the gamble button and the event where you move into a space station.

Note to self to find the best way to hide the gamble button in the case that the user doesn't click it and just moves to the next CP. Maybe through a listener for a location change?