Fixed bug of devMenu population of non-selected items at 0,0
A bug was occurring where if you opened the devMenu on start and only changed the starting location all non-selected items would be loaded to point 0,0 causing a immediate meteor collision even if you didn't move any distance.

The issue was when pulling the data out of localstroage and converting strings to ints I was using parseInt instead of parseFloat thus setting all the 0.5 values that were used to mark the items not to populate on the map to 0's instead of 0.5. Bypassing the if conditional to not populate them on the map. This in turn placed all those items at point 0,0. 

The fix was simply to change the parseInt to parseFloat to keep the 0.5 and be blocked by the if conditional on the map population function.