Bug update for map population at 0,0 for unwanted items.
Bug was due to using parseInt after pulling from localstorage which would change the 0.5 key for unwanted items from the dev menu to 0's which would bypass the check to not populate the map. 

changed the parseInt to parseFloat to keep the 0.5 value when pulling off localstorage and have it not populate onto the map due to the if conditional.