Minor fixes to Casinian gamble
Alert when gambling with Casinian states if you take more than the Casinian it will take double from what you tried to take, however that wasn't implemented. 
Added a simple 2*amountTakenByUser onto both the alert and the arithmetic expression to apply that req.
Also added a more decriptive statement for when the Casinsian is available since it was easy to miss that the Gamble button was popping up below the sensors button to begin the event.